third_party/WebKit/LayoutTests/css3/device-adapt/viewport-invalid-values-001.html

<!DOCTYPE html>
<html>
<head>
    <title>'initial' and 'inherit' are invalid values for @viewport descriptors</title>
    <script src="../../resources/testharness.js"></script>
    <script src="../../resources/testharnessreport.js"></script>
    <script>
        if (window.testRunner) {
            internals.settings.setViewportEnabled(true);
        }
    </script>
    <style>
        /* 'initial' and 'inherit' are invalid keywords for viewport descriptors. */
        @viewport { min-width: initial; }
        @viewport { min-width: inherit; }
    </style>
</head>
<body>
    <div id="log"></div>
    <script>
        test(function(){
            assert_equals(document.styleSheets[0].cssRules.length, 2);
        }, "Check that both @viewport rules are created.");
        test(function(){
            assert_equals(document.styleSheets[0].cssRules[0].style.length, 0);
        }, "Check that \"min-width: initial\" is dropped.");
        test(function(){
            assert_equals(document.styleSheets[0].cssRules[0].style.length, 0);
        }, "Check that \"min-width: inherit\" is dropped.");
    </script>
</body>
</html>
